Version 1.1 of remove-RGPolicyAssignment.ps1

+++ Remove-RGPolicyAssignment.ps1
@@ -25,41 +25,82 @@
     .NOTES
         Author: Simon Baker
         Created: 2018-11-02
-        Modified: 
-        Version: 1.0
+        Modified: 2018-11-02
+        Version: 1.1
+
+        Change Log:
+            1.1 - Added Try-catch to address mid operation failures
 #>
 
 [CmdletBinding(SupportsShouldProcess=$True)]
 Param(
     [parameter(Mandatory=$true,Position=1)][string]$ResourceGroup,
     [parameter()][string]$subscriptionID,
-    [parameter()][switch]$auto
+    [parameter()][switch]$auto,
+    [parameter()][switch]$removeallcustdef
 )
 
+# Variable definitions
+[array]$definitions=@()
 if(!($subscriptionID)){
     Write-Verbose "No subcription id provided, using current context..."
     $subscriptionID = (Get-AzureRmContext).Subscription
     Write-Verbose "Subscription ID: $($subscriptionID)"
 }
+$scopeID = "/subscriptions/" + $subscriptionID + "/resourceGroups/" + $ResourceGroup 
 
-$scopeID = "/subscriptions/" + $subscriptionID + "/resourceGroups/" + $ResourceGroup
-[array]$definitions=@() 
-
+# Main script start
+# Removing Policy Assignments
 Write-Verbose "Getting policy assignments..."
 $policyassignments = Get-AzureRmPolicyAssignment -scope $scopeID
 Write-Host "$($policyassignments.Count) policies found"
 if(!$auto){pause}
-$policyassignments | ForEach-Object {
-    Write-Verbose "Removing Assignment: $($_.Name)"
-    $definitions += $_.Properties.policyDefinitionId
-    Remove-AzureRmPolicyAssignment -Id $_.PolicyAssignmentId
+try{
+    $policyassignments | ForEach-Object {
+        Write-Verbose "Removing Assignment: $($_.Name)"
+        Remove-AzureRmPolicyAssignment -Id $_.PolicyAssignmentId
+        $definitions += $_.Properties.policyDefinitionId
+    }
+}catch{
+    Write-Error $_
+    if($definitions){
+        Write-Error "Policy Definitions may be left behind due to script failing before definition step. Definitions below may need manual removal"
+        Write-Error $definitions
+    }
+    Exit 10
 }
 
-Write-Verbose "Getting Custom Definitions..."
+# Removing Policy Definitions used by the removed policy assignments
+Write-Verbose "Getting Custom Definitions used in removed assignments..."
 $customdefinitions = Get-AzureRmPolicyDefinition | Where-Object {($_.Properties.policyType -like "Custom") -and ($definitions -contains $_.ResourceId)}
 Write-Host "$($customdefinitions.count) custom definitions found"
 if(!$auto){pause}
-$customdefinitions | foreach-object {
-    Write-Verbose "Removing Definition: $($_.Properties.displayName)"
-    Remove-AzureRmPolicyDefinition -Id $_.PolicyDefinitionId -Force
+try{
+    $customdefinitions | foreach-object {
+        Write-Verbose "Removing Definition: $($_.Properties.displayName)"
+        Remove-AzureRmPolicyDefinition -Id $_.PolicyDefinitionId -Force
+    }
+}catch{
+    Write-Error $_
+    Write-Error "Error during definition removal, not all defintions may have been removed."
+    Write-Error "Manual removal may be required for the following definitions:"
+    $customdefinitions.PolicyDefinitionId | Write-Error
+    Exit 20
+}
+
+# attempt to remove all custom definitions
+# no try catch here so script will report error for each definition that fails
+Write-Verbose "Checking for remaining definitions..."
+$remainingDefs = Get-AzureRmPolicyDefinition | Where-Object {($_.Properties.policyType -like "Custom")}
+if($remainingDefs){
+    Write-Host "$($remainingDefs.count) remaining custom definitions"
+    if(!$auto){$continue = Read-Host "Continue with removal? [yes/no]"}else{$continue = "yes"}
+    if($continue -eq "yes"){
+        $remainingDefs | ForEach-Object {
+            Write-Verbose "Removing Definition: $($_.Properties.displayName)"
+            Remove-AzureRmPolicyDefinition -Id $_.PolicyDefinitionId -Force
+        }
+    }else{
+        Write-Verbose "Canceling removal..."
+    }
 }
